Implement qwalk and specdecomp in qtools.js
The numeric.eig function in Numeric.js is unable to compute the
eigenvectors for the adjacency matrix
[[0 1 0]
 [1 0 1]
 [0 1 0]]
because the solution fails to converge, even if the maximum number of
iterations or the tolerance is increased. Therefore, this commit adds
the compRealEigSys.js script, which computes the eigenvectors of the
adjacency matrix.

This commit also adds implementations of the specdecomp and qwalk
functions. Now we can compute the quantum walk. Note that the
implementations are untested.