Finished implementing back-end and basic animation
This commit fully implements the back-end that is responsible for
computing the quantum walk probabilities. It also adds a basic animation
that allows a user to visualize the quantum walk.

The js/initGraph.js script is deleted because the initGraph function is
incorporated into js/main.js.

The js/valToColor.js script is deleted because the valToColor function
is incorporated into js/qwalk.js.

The lib/Numeric.js file is deleted because it is a duplicate of
lib/numeric.js.

The index.html file is modified to import each script individually,
rather than importing a pair of *combined.js files. This is done to
simplify testing.

The js/main.js script is modified to link the background color of each
node with its 'bg' property. Also, a basic animation is set up and run
from the main function.

The js/qtools.js script is modified to properly implement the back-end
of the quantum walk computations, including the spectral decomposition
function and the time-based qwalk function.

The js/qwalk.js script is modified to include an init function that
generates a simple graph and a step function that computes a single time
step in the quantum walk.