Re-implement quantum walk not to use cy.animation
This commit re-implements the quantum walk animation so that it relies
on setInterval() and clearInterval() rather than Cytoscape's animation
functionality. This simplifies starting and stopping the animation, and
will later simplify pausing functionality when it is added to the
existing framework.

js/main.js is modified to add a black border around the nodes, except
the start node, which has a bluish border as an identifier. Also, the
inital graph is constructed directly, rather than through a series of
function calls.

js/qmanip.js is trimmed down to include only what is necessary. Since
qwalk.startFromGraph() uses a breadth first search and the start node to
construct the adjacency matrix, the functionality of the node id table
in js/qmanip.js is unnecessary.

The function qtools.qwalk() in js/qtools.js is modified to take as input
the spectral decomposition matrix, which is derived from the adjacency
matrix and stored in js/qwalk.js; the number of nodes in the adjacency
matrix, which is determined using the breadth first search algorithm;
and the time parameter, which is used to compute the probability
amplitudes.

js/qwalk.js is also trimmed to eliminate unused functions, since we are
no longer using Cytoscape's animation functionality. The startFromGraph
function is rewritten to use the breadth first search algorithm and the
start node to generate the adjacency matrix that is later used in the
quantum walk calculations. The loop function is slightly modified to
reflect the fact that the spectral decomposition matrix is being stored
and not recomputed.

Finally, js/GUI.js is modified to start and stop the quantum walk
animation by calling qwalk.startFromGraph() when the Play button is
clicked and calling clearInterval() when the Stop button is clicked.