Add back-end support to select start node
This commit adds a new function to js/qmanip.js to specify the starting
node using its id.

This commit also re-implements qwalk.startFromGraph() to generate the
adjacency matrix by simply iterating over cy.nodes(), rather than using
the breadth first search algorithm. This chance is made for simplicity,
since the order of the nodes in the matrix is irrelavant. Also, to get
the correct amplitudes for the probability distribution in qwalk.loop(),
we simply look at the column in the amplitude matrix (U) corresponding
to the start index.

Note: this commit adds code to allow a start node to be selected within
the program, but code needs to be added to allow the user to select the
start node through the UI.