Implement bug fix for backend and add testing code
This commit fixes a subtle bug in the backend that causes the quantum
walk animation to behave abnormally for certain graph configurations.
The bug occurred when the eigenvectors returned by qtools.eig() were not
orthogonal. To correct this, the function qtools.MGS() (short for
Modified Gram-Schmidt) is added to orthogonalize the eigenvectors.

Also, to verify that all the backend code is computationally correct,
several test functions were added to js/qtools.js. The function
qtools.testAll() runs through all of the tests in order. The parameter
qtools.tolerance specifies the minimul level of accuracy required for
the test functions to pass successfully. It is currently set to require
15 decimal places of accuracy.

The code has already been tested, and all tests pass successfully.

js/main.js simply adds a call to qtools.testAll() so that all tests are
run at the time that the page is first loaded.

js/qwalk.js is modified to reflect the fact that most of the code in
js/qtools.js no longer uses tensor types, as this makes the code slower
and more verbose. Tensor types (created using numeric.t()) are not used
except in qtools.qwalk(), so qtools.eig() and qtools.specdecomp() do not
use or return them. Instead, the results of qtools.specdecomp() are
converted to tensor types and cached in js/qwalk.js. This change is made
for two reasons. First, the adjacency matrix for an undirected graph is
always symmetric, so the eigenvalues and eigenvectors are always real.
This makes using tensor types unnecessary for testing purposes. Tensors
are needed for qtools.qwalk(), which is why they are created and cached
in js/qwalk.js.